Move TestHTTPServer to its own file
